fix(web): dependencies
