+++ distribution/images/builder-nodejs/Dockerfile
@@ -1,8 +1,8 @@
-FROM docker.io/library/node:20-alpine3.17
+FROM docker.io/library/node:24-alpine3.22
 
 SHELL ["/bin/ash","-c"]
 WORKDIR /usr/work
 LABEL org.opencontainers.image.source="https://github.com/dyrector-io/dyrectorio/tree/develop/distribution/images/builder-nodejs"
 
-RUN apk add alpine-sdk docker zstd openssl1.1-compat jq
+RUN apk add alpine-sdk docker zstd jq
