+++ web/crux/package.json
@@ -40,53 +40,52 @@
   "dependencies": {
     "@grpc/grpc-js": "^1.9.15",
     "@grpc/proto-loader": "^0.7.8",
-    "@nestjs-modules/mailer": "^1.9.1",
-    "@nestjs/axios": "^3.0.0",
-    "@nestjs/cache-manager": "^2.3.0",
-    "@nestjs/cli": "^10.4.5",
-    "@nestjs/common": "^10.4.16",
-    "@nestjs/config": "^3.0.0",
-    "@nestjs/core": "^10.4.2",
-    "@nestjs/event-emitter": "^2.0.3",
-    "@nestjs/jwt": "^10.1.0",
-    "@nestjs/microservices": "^10.1.3",
-    "@nestjs/passport": "^10.0.0",
-    "@nestjs/platform-express": "^10.4.20",
-    "@nestjs/platform-ws": "^10.1.3",
-    "@nestjs/swagger": "^7.4.1",
-    "@nestjs/websockets": "^10.1.3",
+    "@nestjs/axios": "^4.0.1",
+    "@nestjs/cache-manager": "^3.0.1",
+    "@nestjs/cli": "^11.0.14",
+    "@nestjs/common": "^11.1.9",
+    "@nestjs/config": "^4.0.2",
+    "@nestjs/core": "^11.1.9",
+    "@nestjs/event-emitter": "^3.0.1",
+    "@nestjs/jwt": "^11.0.2",
+    "@nestjs/microservices": "^11.1.9",
+    "@nestjs/passport": "^11.0.5",
+    "@nestjs/platform-express": "^11.1.9",
+    "@nestjs/platform-ws": "^11.1.9",
+    "@nestjs/swagger": "^11.2.3",
+    "@nestjs/websockets": "^11.1.9",
     "@ory/kratos-client": "^1.2.1",
     "@prisma/client": "5.1.0",
     "@types/google-protobuf": "^3.15.6",
     "@types/semver": "^7.5.0",
     "@types/ws": "^8.5.5",
-    "@willsoto/nestjs-prometheus": "^5.2.1",
-    "axios": "^1.12.2",
-    "cache-manager": "^5.7.6",
+    "@willsoto/nestjs-prometheus": "^6.0.2",
+    "axios": "^1.13.2",
+    "cache-manager": "^7.2.7",
     "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
-    "dotenv": "^16.3.1",
+    "class-validator": "^0.14.3",
+    "dotenv": "^17.2.3",
     "google-auth-library": "^9.0.0",
     "google-protobuf": "^3.21.2",
     "handlebars": "^4.7.8",
-    "nestjs-pino": "^3.3.0",
-    "nodemailer": "^6.9.9",
-    "passport": "^0.6.0",
+    "nestjs-pino": "^4.5.0",
+    "nodemailer": "^7.0.11",
+    "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "pg": "^8.11.2",
-    "pino-http": "^8.3.3",
-    "prom-client": "^14.2.0",
+    "pino-http": "^11.0.0",
+    "prom-client": "^15.0.0",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.8.1",
+    "rxjs": "^7.8.2",
     "semver": "^7.5.4",
     "uuid": "^9.0.0",
-    "ws": "^8.13.0",
+    "ws": "^8.18.3",
     "www-authenticate": "^0.6.3",
     "yup": "^1.2.0"
   },
   "devDependencies": {
-    "@nestjs/schematics": "^10.0.1",
-    "@nestjs/testing": "^10.1.3",
+    "@nestjs/schematics": "^11.0.9",
+    "@nestjs/testing": "^11.1.9",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.3",
     "@types/node": "^20.4.5",
@@ -102,7 +101,7 @@
     "eslint-config-prettier": "^8.9.0",
     "eslint-plugin-import": "^2.28.0",
     "jest": "^29.6.2",
-    "pino-pretty": "^10.2.0",
+    "pino-pretty": "^13.1.3",
     "prettier": "^3.0.0",
     "prettier-plugin-organize-imports": "^3.2.3",
     "prisma": "5.1.0",
@@ -110,12 +109,6 @@
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.1.6"
-  },
-  "overrides": {
-    "degenerator": "5.0.0",
-    "@nestjs-modules/mailer": {
-      "mjml": "^5.0.0-alpha.4"
-    }
+    "typescript": "^5.9.3"
   }
 }
