+++ web/crux/src/mailer/email.module.ts
@@ -1,38 +1,8 @@
-import { MailerModule } from '@nestjs-modules/mailer'
-import { TransportType } from '@nestjs-modules/mailer/dist/interfaces/mailer-options.interface'
 import { Module } from '@nestjs/common'
-import { ConfigModule, ConfigService } from '@nestjs/config'
-import { parseConnectionUrl } from 'nodemailer/lib/shared'
 import EmailService from './email.service'
 
-const smtpUriToTransport = (uri: string): TransportType => {
-  const transport = parseConnectionUrl(uri)
-
-  if (uri.toLowerCase().includes('skip_ssl_verify=true')) {
-    transport.tls = {
-      rejectUnauthorized: false,
-    }
-  }
-
-  return transport
-}
-
 @Module({
-  imports: [
-    MailerModule.forRootAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: async (configService: ConfigService) => ({
-        transport: smtpUriToTransport(configService.getOrThrow('SMTP_URI')),
-        defaults: {
-          from: {
-            address: configService.get('FROM_EMAIL'),
-            name: configService.get('FROM_NAME'),
-          },
-        },
-      }),
-    }),
-  ],
+  imports: [],
   controllers: [],
   providers: [EmailService],
 })
